Fascia_dataViz/Fascia_sensor_data_plotter/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
ca-certificates=2019.8.28=0
certifi=2019.9.11=py37_0
dbus=1.13.6=h90a0687_0
expat=2.2.6=h0a44026_0
gettext=0.19.8.1=h15daf44_3
glib=2.56.2=hd9629dc_0
icu=58.2=h4b95b61_1
intel-openmp=2019.4=233
jpeg=9b=he5867d9_2
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
libgfortran=3.0.1=h93005f0_2
libiconv=1.15=hdd342a3_7
libpng=1.6.37=ha441bb4_0
mkl=2019.4=233
mkl-service=2.3.0=py37hfbe908c_0
mkl_fft=1.0.14=py37h5e564d8_0
mkl_random=1.1.0=py37ha771720_0
ncurses=6.1=h0a44026_1
numpy=1.17.2=py37h99e6662_0
numpy-base=1.17.2=py37h6575580_0
openssl=1.1.1d=h1de35cc_2
pandas=0.25.1=py37h0a44026_0
pcre=8.43=h0a44026_0
pip=19.2.3=py37_0
pyqt=5.9.2=py37h655552a_2
pyqtgraph=0.10.0=py37h28b3542_3
python=3.7.4=h359304d_1
python-dateutil=2.8.0=py37_0
pytz=2019.3=py_0
qt=5.9.7=h468cd18_1
readline=7.0=h1de35cc_5
setuptools=41.4.0=py37_0
sip=4.19.8=py37h0a44026_0
six=1.12.0=py37_0
sqlite=3.30.0=ha441bb4_0
tk=8.6.8=ha441bb4_0
wheel=0.33.6=py37_0
xz=5.2.4=h1de35cc_4
zlib=1.2.11=h1de35cc_3
